[수정] 캐스팅

+++ plugins/err-ruli/ruli.py
@@ -26,7 +26,7 @@ class Ruli(BotPlugin):
         send_content = ''
         for best in range(1, 31):
             if best % 10 == 1:
-                send_content += '[' + main_title[best / 10] + ']\n'
+                send_content += '[' + main_title[(int)(best / 10)] + ']\n'
 
             link = context[context.find(item_start_tag):context.find(item_end_tag)]
             sub_title = context[context.find(item_end_tag) + len(item_end_tag):context.find(title_end_tag)]
